+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 - [f6e15db] [Feature] Add Gemini 2.5 Flash
 - [3dabeac] [Feature] Add o3 and o4-mini
 
+> Improvements:
+- [d02c25a] [Improvement] Add reasoning to AI managed mode
+
 > Bug Fixes:
 - [62c451b] [BugFix] Get reasoning_effort value properly
 - [85ba311] [BugFix] get o4 models properly
-- [5ccff7c] [BugFix] Typo
 
 > Others:
 - [833b7ec] [Others] Remove unused import